Regroup all rcp submodules inside rcp module.
