+++ projects/VisualStudio2013/mupen64plus-core.vcxproj
@@ -49,7 +49,7 @@
     <ClCompile Include="..\..\src\debugger\dbg_debugger.c" />
     <ClCompile Include="..\..\src\debugger\dbg_decoder.c" />
     <ClCompile Include="..\..\src\debugger\dbg_memory.c" />
-    <ClCompile Include="..\..\src\device\ai\ai_controller.c" />
+    <ClCompile Include="..\..\src\device\rcp\ai\ai_controller.c" />
     <ClCompile Include="..\..\src\device\cart\af_rtc.c" />
     <ClCompile Include="..\..\src\device\cart\cart.c" />
     <ClCompile Include="..\..\src\device\cart\cart_rom.c" />
@@ -105,7 +105,7 @@
     <ClCompile Include="..\..\src\osd\OGLFT.cpp" />
     <ClCompile Include="..\..\src\osd\osd.cpp" />
     <ClCompile Include="..\..\src\osd\screenshot.cpp" />
-    <ClCompile Include="..\..\src\device\pi\pi_controller.c" />
+    <ClCompile Include="..\..\src\device\rcp\pi\pi_controller.c" />
     <ClCompile Include="..\..\src\plugin\dummy_audio.c" />
     <ClCompile Include="..\..\src\plugin\dummy_input.c" />
     <ClCompile Include="..\..\src\plugin\dummy_rsp.c" />
@@ -127,7 +127,7 @@
     <ClCompile Include="..\..\src\device\r4300\exception.c" />
     <ClCompile Include="..\..\src\device\r4300\instr_counters.c" />
     <ClCompile Include="..\..\src\device\r4300\interrupt.c" />
-    <ClCompile Include="..\..\src\device\mi\mi_controller.c" />
+    <ClCompile Include="..\..\src\device\rcp\mi\mi_controller.c" />
     <ClCompile Include="..\..\src\device\r4300\new_dynarec\arm\arm_cpu_features.c">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">true</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">true</ExcludedFromBuild>
@@ -230,17 +230,17 @@
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='New_Dynarec_Release|Win32'">true</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='New_Dynarec_Release|x64'">true</ExcludedFromBuild>
     </ClCompile>
-    <ClCompile Include="..\..\src\device\rdp\fb.c" />
-    <ClCompile Include="..\..\src\device\rdp\rdp_core.c" />
-    <ClCompile Include="..\..\src\device\ri\rdram.c" />
-    <ClCompile Include="..\..\src\device\ri\rdram_detection_hack.c" />
-    <ClCompile Include="..\..\src\device\ri\ri_controller.c" />
-    <ClCompile Include="..\..\src\device\rsp\rsp_core.c" />
+    <ClCompile Include="..\..\src\device\rcp\rdp\fb.c" />
+    <ClCompile Include="..\..\src\device\rcp\rdp\rdp_core.c" />
+    <ClCompile Include="..\..\src\device\rcp\ri\rdram.c" />
+    <ClCompile Include="..\..\src\device\rcp\ri\rdram_detection_hack.c" />
+    <ClCompile Include="..\..\src\device\rcp\ri\ri_controller.c" />
+    <ClCompile Include="..\..\src\device\rcp\rsp\rsp_core.c" />
     <ClCompile Include="..\..\src\device\pif\cic.c" />
     <ClCompile Include="..\..\src\device\pif\n64_cic_nus_6105.c" />
     <ClCompile Include="..\..\src\device\pif\pif.c" />
-    <ClCompile Include="..\..\src\device\si\si_controller.c" />
-    <ClCompile Include="..\..\src\device\vi\vi_controller.c" />
+    <ClCompile Include="..\..\src\device\rcp\si\si_controller.c" />
+    <ClCompile Include="..\..\src\device\rcp\vi\vi_controller.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\src\api\callbacks.h" />
@@ -270,7 +270,7 @@
     <ClInclude Include="..\..\src\debugger\dbg_decoder_local.h" />
     <ClInclude Include="..\..\src\debugger\dbg_memory.h" />
     <ClInclude Include="..\..\src\debugger\dbg_types.h" />
-    <ClInclude Include="..\..\src\device\ai\ai_controller.h" />
+    <ClInclude Include="..\..\src\device\rcp\ai\ai_controller.h" />
     <ClInclude Include="..\..\src\device\cart\af_rtc.h" />
     <ClInclude Include="..\..\src\device\cart\cart.h" />
     <ClInclude Include="..\..\src\device\cart\cart_rom.h" />
@@ -310,7 +310,7 @@
     <ClInclude Include="..\..\src\osd\OGLFT.h" />
     <ClInclude Include="..\..\src\osd\osd.h" />
     <ClInclude Include="..\..\src\osd\screenshot.h" />
-    <ClInclude Include="..\..\src\device\pi\pi_controller.h" />
+    <ClInclude Include="..\..\src\device\rcp\pi\pi_controller.h" />
     <ClInclude Include="..\..\src\plugin\dummy_audio.h" />
     <ClInclude Include="..\..\src\plugin\dummy_input.h" />
     <ClInclude Include="..\..\src\plugin\dummy_rsp.h" />
@@ -324,7 +324,7 @@
     <ClInclude Include="..\..\src\device\r4300\instr_counters.h" />
     <ClInclude Include="..\..\src\device\r4300\interrupt.h" />
     <ClInclude Include="..\..\src\device\r4300\macros.h" />
-    <ClInclude Include="..\..\src\device\mi\mi_controller.h" />
+    <ClInclude Include="..\..\src\device\rcp\mi\mi_controller.h" />
     <ClInclude Include="..\..\src\device\r4300\new_dynarec\arm\arm_cpu_features.h">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">true</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">true</ExcludedFromBuild>
@@ -444,17 +444,17 @@
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='New_Dynarec_Release|Win32'">true</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='New_Dynarec_Release|x64'">true</ExcludedFromBuild>
     </ClInclude>
-    <ClInclude Include="..\..\src\device\rdp\fb.h" />
-    <ClInclude Include="..\..\src\device\rdp\rdp_core.h" />
-    <ClInclude Include="..\..\src\device\ri\rdram.h" />
-    <ClInclude Include="..\..\src\device\ri\rdram_detection_hack.h" />
-    <ClInclude Include="..\..\src\device\ri\ri_controller.h" />
-    <ClInclude Include="..\..\src\device\rsp\rsp_core.h" />
+    <ClInclude Include="..\..\src\device\rcp\rdp\fb.h" />
+    <ClInclude Include="..\..\src\device\rcp\rdp\rdp_core.h" />
+    <ClInclude Include="..\..\src\device\rcp\ri\rdram.h" />
+    <ClInclude Include="..\..\src\device\rcp\ri\rdram_detection_hack.h" />
+    <ClInclude Include="..\..\src\device\rcp\ri\ri_controller.h" />
+    <ClInclude Include="..\..\src\device\rcp\rsp\rsp_core.h" />
     <ClInclude Include="..\..\src\device\pif\cic.h" />
     <ClInclude Include="..\..\src\device\pif\n64_cic_nus_6105.h" />
     <ClInclude Include="..\..\src\device\pif\pif.h" />
-    <ClInclude Include="..\..\src\device\si\si_controller.h" />
-    <ClInclude Include="..\..\src\device\vi\vi_controller.h" />
+    <ClInclude Include="..\..\src\device\rcp\si\si_controller.h" />
+    <ClInclude Include="..\..\src\device\rcp\vi\vi_controller.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\..\src\device\r4300\mips_instructions.def" />
