+++ src/asm_defines/asm_defines.c
@@ -33,7 +33,7 @@
 #include "device/memory/memory.h"
 #include "device/r4300/new_dynarec/new_dynarec.h"
 #include "device/r4300/r4300_core.h"
-#include "device/ri/ri_controller.h"
+#include "device/rcp/ri/ri_controller.h"
 
 #include <stddef.h>
