+++ src/debugger/dbg_memory.c
@@ -27,17 +27,17 @@
 #include "dbg_breakpoints.h"
 #include "dbg_memory.h"
 #include "dbg_types.h"
-#include "device/ai/ai_controller.h"
 #include "device/memory/memory.h"
-#include "device/mi/mi_controller.h"
-#include "device/pi/pi_controller.h"
 #include "device/pif/pif.h"
 #include "device/r4300/r4300_core.h"
-#include "device/rdp/rdp_core.h"
-#include "device/ri/ri_controller.h"
-#include "device/rsp/rsp_core.h"
-#include "device/si/si_controller.h"
-#include "device/vi/vi_controller.h"
+#include "device/rcp/ai/ai_controller.h"
+#include "device/rcp/mi/mi_controller.h"
+#include "device/rcp/pi/pi_controller.h"
+#include "device/rcp/rdp/rdp_core.h"
+#include "device/rcp/ri/ri_controller.h"
+#include "device/rcp/rsp/rsp_core.h"
+#include "device/rcp/si/si_controller.h"
+#include "device/rcp/vi/vi_controller.h"
 #include "main/main.h"
 
 #if !defined(NO_ASM) && (defined(__i386__) || (defined(__x86_64__) && defined(__GNUC__)))
