+++ src/device/cart/cart.c
@@ -25,7 +25,7 @@
 #include "api/m64p_types.h"
 
 #include "main/rom.h"
-#include "device/ri/rdram.h"
+#include "device/rcp/ri/rdram.h"
 
 #include <stdint.h>
 #include <string.h>
