+++ src/device/device.h
@@ -25,23 +25,23 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#include "ai/ai_controller.h"
-#include "memory/memory.h"
-#include "mi/mi_controller.h"
-#include "pi/pi_controller.h"
-#include "pif/pif.h"
-#include "r4300/r4300_core.h"
-#include "rdp/rdp_core.h"
-#include "ri/ri_controller.h"
-#include "rsp/rsp_core.h"
-#include "si/si_controller.h"
-#include "vi/vi_controller.h"
+#include "cart/cart.h"
 #include "controllers/game_controller.h"
 #include "controllers/paks/mempak.h"
 #include "controllers/paks/rumblepak.h"
 #include "controllers/paks/transferpak.h"
-#include "cart/cart.h"
 #include "gb/gb_cart.h"
+#include "memory/memory.h"
+#include "pif/pif.h"
+#include "r4300/r4300_core.h"
+#include "rcp/ai/ai_controller.h"
+#include "rcp/mi/mi_controller.h"
+#include "rcp/pi/pi_controller.h"
+#include "rcp/rdp/rdp_core.h"
+#include "rcp/ri/ri_controller.h"
+#include "rcp/rsp/rsp_core.h"
+#include "rcp/si/si_controller.h"
+#include "rcp/vi/vi_controller.h"
 
 struct audio_out_backend_interface;
 struct storage_backend_interface;
