+++ src/device/memory/memory.c
@@ -25,7 +25,7 @@
 #include "api/m64p_types.h"
 
 #include "device/device.h"
-#include "device/rsp/rsp_core.h"
+#include "device/rcp/rsp/rsp_core.h"
 #include "device/pif/pif.h"
 
 #ifdef DBG
