+++ src/device/rcp/ai/ai_controller.c
@@ -25,10 +25,10 @@
 
 #include "backends/api/audio_out_backend.h"
 #include "device/memory/memory.h"
-#include "device/mi/mi_controller.h"
 #include "device/r4300/r4300_core.h"
-#include "device/ri/ri_controller.h"
-#include "device/vi/vi_controller.h"
+#include "device/rcp/mi/mi_controller.h"
+#include "device/rcp/ri/ri_controller.h"
+#include "device/rcp/vi/vi_controller.h"
 
 enum
 {
