+++ src/device/rcp/mi/mi_controller.c
