+++ src/device/rcp/mi/mi_controller.h
@@ -1,5 +1,5 @@
 /* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
- *   Mupen64plus - ri_controller.h                                         *
+ *   Mupen64plus - mi_controller.h                                         *
  *   Mupen64Plus homepage: http://code.google.com/p/mupen64plus/           *
  *   Copyright (C) 2014 Bobby Smiles                                       *
  *                                                                         *
@@ -19,46 +19,53 @@
  *   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.          *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-#ifndef M64P_DEVICE_RI_RI_CONTROLLER_H
-#define M64P_DEVICE_RI_RI_CONTROLLER_H
+#ifndef M64P_DEVICE_RCP_MI_MI_CONTROLLER_H
+#define M64P_DEVICE_RCP_MI_MI_CONTROLLER_H
 
-#include <stddef.h>
 #include <stdint.h>
 
-#include "rdram.h"
+struct r4300_core;
 
-enum ri_registers
+enum mi_registers
 {
-    RI_MODE_REG,
-    RI_CONFIG_REG,
-    RI_CURRENT_LOAD_REG,
-    RI_SELECT_REG,
-    RI_REFRESH_REG,
-    RI_LATENCY_REG,
-    RI_ERROR_REG,
-    RI_WERROR_REG,
-    RI_REGS_COUNT
+    MI_INIT_MODE_REG,
+    MI_VERSION_REG,
+    MI_INTR_REG,
+    MI_INTR_MASK_REG,
+    MI_REGS_COUNT
 };
 
-struct ri_controller
-{
-    uint32_t regs[RI_REGS_COUNT];
 
-    struct rdram rdram;
+enum mi_intr
+{
+    MI_INTR_SP = 0x01,
+    MI_INTR_SI = 0x02,
+    MI_INTR_AI = 0x04,
+    MI_INTR_VI = 0x08,
+    MI_INTR_PI = 0x10,
+    MI_INTR_DP = 0x20
 };
 
-static uint32_t ri_reg(uint32_t address)
+struct mi_controller
+{
+    uint32_t regs[MI_REGS_COUNT];
+
+    struct r4300_core* r4300;
+};
+
+static uint32_t mi_reg(uint32_t address)
 {
     return (address & 0xffff) >> 2;
 }
 
-void init_ri(struct ri_controller* ri,
-             uint32_t* dram,
-             size_t dram_size);
+void init_mi(struct mi_controller* mi, struct r4300_core* r4300);
+void poweron_mi(struct mi_controller* mi);
 
-void poweron_ri(struct ri_controller* ri);
+void read_mi_regs(void* opaque, uint32_t address, uint32_t* value);
+void write_mi_regs(void* opaque, uint32_t address, uint32_t value, uint32_t mask);
 
-void read_ri_regs(void* opaque, uint32_t address, uint32_t* value);
-void write_ri_regs(void* opaque, uint32_t address, uint32_t value, uint32_t mask);
+void raise_rcp_interrupt(struct mi_controller* mi, uint32_t mi_intr);
+void signal_rcp_interrupt(struct mi_controller* mi, uint32_t mi_intr);
+void clear_rcp_interrupt(struct mi_controller* mi, uint32_t mi_intr);
 
 #endif
