+++ src/device/rcp/ri/rdram.h
@@ -1,5 +1,5 @@
 /* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
- *   Mupen64plus - rdram_detection_hack.h                                  *
+ *   Mupen64plus - rdram.h                                                 *
  *   Mupen64Plus homepage: http://code.google.com/p/mupen64plus/           *
  *   Copyright (C) 2014 Bobby Smiles                                       *
  *                                                                         *
@@ -19,12 +19,54 @@
  *   51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.          *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-#ifndef M64P_DEVICE_RI_RDRAM_DETECTION_HACK_H
-#define M64P_DEVICE_RI_RDRAM_DETECTION_HACK_H
+#ifndef M64P_DEVICE_RCP_RI_RDRAM_H
+#define M64P_DEVICE_RCP_RI_RDRAM_H
 
-struct rdram;
-struct cic;
+#include <stddef.h>
+#include <stdint.h>
 
-void force_detected_rdram_size_hack(struct rdram* rdram, const struct cic* cic);
+enum rdram_registers
+{
+    RDRAM_CONFIG_REG,
+    RDRAM_DEVICE_ID_REG,
+    RDRAM_DELAY_REG,
+    RDRAM_MODE_REG,
+    RDRAM_REF_INTERVAL_REG,
+    RDRAM_REF_ROW_REG,
+    RDRAM_RAS_INTERVAL_REG,
+    RDRAM_MIN_INTERVAL_REG,
+    RDRAM_ADDR_SELECT_REG,
+    RDRAM_DEVICE_MANUF_REG,
+    RDRAM_REGS_COUNT
+};
+
+struct rdram
+{
+    uint32_t regs[RDRAM_REGS_COUNT];
+    uint32_t* dram;
+    size_t dram_size;
+};
+
+static uint32_t rdram_reg(uint32_t address)
+{
+    return (address & 0x3ff) >> 2;
+}
+
+static uint32_t rdram_dram_address(uint32_t address)
+{
+    return (address & 0xffffff) >> 2;
+}
+
+void init_rdram(struct rdram* rdram,
+                uint32_t* dram,
+                size_t dram_size);
+
+void poweron_rdram(struct rdram* rdram);
+
+void read_rdram_regs(void* opaque, uint32_t address, uint32_t* value);
+void write_rdram_regs(void* opaque, uint32_t address, uint32_t value, uint32_t mask);
+
+void read_rdram_dram(void* opaque, uint32_t address, uint32_t* value);
+void write_rdram_dram(void* opaque, uint32_t address, uint32_t value, uint32_t mask);
 
 #endif
