+++ src/device/rcp/ri/ri_controller.c
