+++ src/device/rcp/rsp/rsp_core.c
@@ -24,10 +24,10 @@
 #include <string.h>
 
 #include "device/memory/memory.h"
-#include "device/mi/mi_controller.h"
 #include "device/r4300/r4300_core.h"
-#include "device/rdp/rdp_core.h"
-#include "device/ri/ri_controller.h"
+#include "device/rcp/mi/mi_controller.h"
+#include "device/rcp/rdp/rdp_core.h"
+#include "device/rcp/ri/ri_controller.h"
 #include "main/main.h"
 #include "main/profile.h"
 #include "plugin/plugin.h"
