+++ src/device/rcp/si/si_controller.c
@@ -26,10 +26,10 @@
 #include "api/callbacks.h"
 #include "api/m64p_types.h"
 #include "device/memory/memory.h"
-#include "device/mi/mi_controller.h"
 #include "device/pif/pif.h"
 #include "device/r4300/r4300_core.h"
-#include "device/ri/ri_controller.h"
+#include "device/rcp/mi/mi_controller.h"
+#include "device/rcp/ri/ri_controller.h"
 
 enum
 {
