+++ src/main/savestates.c
@@ -37,17 +37,17 @@
 #include "api/config.h"
 #include "api/m64p_config.h"
 #include "api/m64p_types.h"
-#include "device/ai/ai_controller.h"
 #include "device/memory/memory.h"
-#include "device/pi/pi_controller.h"
 #include "device/r4300/new_dynarec/new_dynarec.h"
 #include "device/r4300/r4300_core.h"
-#include "device/rdp/rdp_core.h"
-#include "device/ri/ri_controller.h"
-#include "device/rsp/rsp_core.h"
-#include "device/si/si_controller.h"
+#include "device/rcp/ai/ai_controller.h"
+#include "device/rcp/pi/pi_controller.h"
+#include "device/rcp/rdp/rdp_core.h"
+#include "device/rcp/ri/ri_controller.h"
+#include "device/rcp/rsp/rsp_core.h"
+#include "device/rcp/si/si_controller.h"
+#include "device/rcp/vi/vi_controller.h"
 #include "device/gb/gb_cart.h"
-#include "device/vi/vi_controller.h"
 #include "main.h"
 #include "main/list.h"
 #include "osal/preproc.h"
