+++ src/plugin/plugin.c
@@ -27,12 +27,12 @@
 #include "api/m64p_common.h"
 #include "api/m64p_plugin.h"
 #include "api/m64p_types.h"
-#include "device/ai/ai_controller.h"
 #include "device/memory/memory.h"
-#include "device/mi/mi_controller.h"
-#include "device/rdp/rdp_core.h"
-#include "device/rsp/rsp_core.h"
-#include "device/vi/vi_controller.h"
+#include "device/rcp/ai/ai_controller.h"
+#include "device/rcp/mi/mi_controller.h"
+#include "device/rcp/rdp/rdp_core.h"
+#include "device/rcp/rsp/rsp_core.h"
+#include "device/rcp/vi/vi_controller.h"
 #include "dummy_audio.h"
 #include "dummy_input.h"
 #include "dummy_rsp.h"
